refactor: 예약 상세 뷰에서  날짜 시간 선택, 영화정보 출력 역할 객체 분리
